project: 24-01-2017: parser-tree = building a Graph: refactored (type ‘int’ - to ‘NSInteger’)

+++ 24-01-2017/parser-tree/parser-tree/parser-tree/composite/Node.h
@@ -18,7 +18,7 @@
 
 
 - (void)insertNode:(Node *)node;
-- (instancetype)initWithData:(int)value;
+- (instancetype)initWithData:(NSInteger)value;
 
 //- (NSString *)printTreev1;
 - (NSString *)printTreev2;
